backend/pyproject.toml
[project]
name = "ai-orchestration-app"
version = "0.1.0"
description = "FastAPI app with LangChain, LangGraph, MongoDB, and Redis"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "motor>=3.3.0",
    "redis>=5.0.0",
    "langchain>=0.1.0",
    "langgraph>=0.0.10",
    "langchain-openai>=0.0.2",
    "langchain-google-genai>=0.0.1",
    "pydantic-settings>=2.1.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "email-validator>=2.0.0",
    "aiohttp>=3.9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
